requirements.txt
flask
gunicorn
azure-data-tables
azure-storage-blob
